# state/runs/2026-10-09-040829-umm-improve/validated-snapshot/docs/openapi.yaml
openapi: 3.1.0
info:
  title: umm API
  version: 0.8.1
  description: |
    Spatial Thought Memory, collaboration, automation and Dream Layer API.
    Supported Canvas mutation requests can include an `Idempotency-Key` header (8-128 safe characters) for replay-safe offline synchronization.
    Personal preferences, API-key lifecycle operations and all administrator operations require an interactive browser session; bearer API keys cannot call them.
    Errors use RFC 9457 Problem Details and note version conflicts include the latest server note.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /spaces/{spaceId}/presentations:
    parameters:
      - { $ref: '#/components/parameters/SpaceID' }
    get:
      operationId: get_spaces_spaceId_presentations
      tags: [Thoughts]
      summary: List the talks made from this space
      x-required-scope: notes:read
      responses:
        '200': { description: Presentation links, content: { application/json: { schema: { type: object, properties: { presentations: { type: array, items: { $ref: '#/components/schemas/PresentationLink' } } } } } } }
    post:
      operationId: post_spaces_spaceId_presentations
      tags: [Thoughts]
      summary: Compile this space into a Ptium deck
      description: >-
        Creates a deck in Ptium and compiles the space into it. Requires
        notes:write rather than notes:read because it sends the author's
        thoughts to another service and creates something there.
      x-required-scope: notes:write
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                noteIds: { type: array, items: { type: string, format: uuid } }
                includeExcluded: { type: boolean }
                oneSlidePerThought: { type: boolean, description: Turn off grouping thoughts by where they sit on the canvas. What the person connected still travels together. }
                nameGroups: { type: boolean, description: "Ask the chat model to name each group of thoughts that were put together by position. Only the headings: the sentences on the slides are the author's either way. Opt-in, because it sends those thoughts to the gateway and makes the deck stop being the same every time. Ignored when no chat model is configured." }
                sectionDeck: { type: boolean, description: "Ask the chat model where a long talk divides into parts. Adds heading slides and moves nothing: the order is a follows chain or the layout the author made, and rearranging it would override what they said. Ignored below 12 slides or with no chat model." }
                maxSlides: { type: integer, minimum: 1, maximum: 500, description: "Cap the talk at this many slides. Omitted or 0 means no cap. Ranking is deterministic and reads only the graph — how many thoughts reached a slide, whether it holds a recorded disagreement — so the same space and the same cap always give the same deck. Part headings count against it too. What did not fit is recorded as trimmedCount." }
      responses:
        '201': { description: The deck, content: { application/json: { schema: { type: object, properties: { link: { $ref: '#/components/schemas/PresentationLink' }, warnings: { type: array, items: { type: string } } } } } } }
        '400': { description: Nothing in this space can become a talk }
        '409': { description: No Ptium is configured }
        '422': { $ref: '#/components/responses/PtiumFailure' }
        '502': { $ref: '#/components/responses/PtiumFailure' }
        '504': { $ref: '#/components/responses/PtiumFailure' }
...
  /notes/{noteId}/presentations:
    parameters:
      - { $ref: '#/components/parameters/NoteID' }
    get:
      operationId: get_notes_noteId_presentations
      tags: [Thoughts]
      summary: Which talks quote this thought
      x-required-scope: notes:read
      responses:
        '200': { description: Presentation links, content: { application/json: { schema: { type: object, properties: { presentations: { type: array, items: { $ref: '#/components/schemas/PresentationLink' } } } } } } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: umm_session }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: umm_key }
  parameters:
    SpaceID: { in: path, name: spaceId, required: true, schema: { type: string, format: uuid } }
    NoteID: { in: path, name: noteId, required: true, schema: { type: string, format: uuid } }
...
  responses:
    PtiumFailure:
      description: >-
        A Ptium call that did not work, sorted by who can do something about it.
        `type` ends in the kind and `failure` repeats it as a bare string, so a
        client can branch without parsing the URL. `detail` is umm's own
        sentence about what to do. `ptiumDetail` is Ptium's own explanation,
        present only when Ptium said something worth repeating — a proxy's HTML
        error page is not passed through. `technical` is the underlying error
        and is sent to administrators only, because it names internal hosts, Go
        types and SQL constraints.
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type: { type: string }
              title: { type: string }
              status: { type: integer }
              detail: { type: string }
              failure: { type: string, enum: [unreachable, timed-out, unauthorized, no-api, rejected, remote-error, unexpected-response, not-recorded, other] }
              ptiumStatus: { type: integer, description: The status Ptium answered with, when it answered. }
              ptiumDetail: { type: string }
              ptiumId: { type: string, description: The deck Ptium opened before compiling failed. Present when a deck was left behind, which is when retrying into it is the right move rather than making another. }
              deckLeftBehind: { type: boolean }
              technical: { type: string, description: Administrators only. }
              requestId: { type: string }
...
  schemas:
...
    PresentationLink:
      type: object
      description: >-
        umm's record that a space produced a deck. Ptium owns the presentation
        and its slides; only the link is kept here, so nothing drifts when the
        deck is edited.
      properties:
        id: { type: string, format: uuid }
        spaceId: { type: string, format: uuid }
        ptiumId: { type: string, description: Ptium's identifier for the deck. Opaque to umm. }
        title: { type: string }
        status: { type: string, enum: [pending, generating, ready, failed] }
        error: { type: string, description: The underlying error, when it failed. Names internal hosts and Go types, so it is for whoever fixes it rather than for display. }
        failureKind: { type: string, description: "What sort of failure it was, which decides the sentence a reader is shown and whether retrying can help. One of unreachable, timed-out, unauthorized, no-api, rejected, remote-error, unexpected-response, not-recorded, other. Empty on decks recorded before v0.60.0 and on decks that did not fail." }
        compiledSource: { type: string, description: The deck source umm sent. Omitted from list responses. }
        thoughtCount: { type: integer }
        excludedCount: { type: integer, description: Thoughts deliberately left out, so the count on screen is the real one. }
        trimmedCount: { type: integer, description: "Thoughts left out only because the talk was capped at a length. Apart from excludedCount because the remedy differs: raising the cap brings these back, and nothing brings back a thought its author held out of analysis. 0 on decks recorded before v0.64.0." }
        createdBy: { type: string, format: uuid, nullable: true }
        url: { type: string, description: Where the deck can be opened. Computed on every read, never stored. Absent when no Ptium is configured. }
        staleSlides:
          type: integer
          description: >-
            How many of this deck's slides quote a thought that has since been
            rewritten or deleted. Moving a thought does not count: the check is
            on the words, not the note's version, which also changes when a note
            is dragged.
...